decided against using footprints with no mask cutouts on top layer
There is no solder mask over the holes and I'm not so sure if that could
mean that the manufacturers will just put the mask over the hole as with
a tented via, which would be bad. So I stick with the exposed pads on
the front which I know work fine.